Stack allocations are now cleared before use.
When allocating from the stack, data within the memory segment is UNDEFINED, meaning it could cause havock if not cleared prior to use. All instantiations of stack allocated memory are now proceeded with a `.Clear()`

On the occasion that stackalloc is used to read data that is immediately thrown away, no `.Clear()` is performed as the data is never actually used any way.